% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_fragments}
\alias{count_fragments}
\title{Count fragments in URLs}
\usage{
count_fragments(url, sort = FALSE, name = "n")
}
\arguments{
\item{url}{A character vector of URLs.}

\item{sort}{Logical indicating whether to sort the output by count. Defaults
to FALSE.}

\item{name}{The name of the column containing the counts. Defaults to 'n'.}
}
\value{
A tibble with each fragment and its count.
}
\description{
Count fragments in URLs
}
\examples{
count_fragments(c("http://example.com#top", "http://example.com#bottom"))
}
