% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_hosts}
\alias{count_hosts}
\title{Count different hosts found in URLs}
\usage{
count_hosts(url, sort = FALSE, name = "n")
}
\arguments{
\item{url}{A character vector of URLs.}

\item{sort}{Logical indicating whether to sort the output by count. Defaults
to FALSE.}

\item{name}{The name of the column containing the counts. Deafults to 'n'.}
}
\value{
A tibble with each host and its count.
}
\description{
Count different hosts found in URLs
}
\examples{
count_hosts(c("http://example.com", "http://www.example.com"))
}
