% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_param_names}
\alias{count_param_names}
\title{Count different parameter names in query strings}
\usage{
count_param_names(query, sort = FALSE, name = "n")
}
\arguments{
\item{query}{A character vector of query strings.}

\item{sort}{Logical indicating whether to sort the output by count. Defaults
to FALSE.}

\item{name}{The name of the column containing the counts. Deafults to 'n'.}
}
\value{
A tibble with each parameter name and how often it occurs.
}
\description{
Count different parameter names in query strings
}
\examples{
count_param_names(c("param1=value1&param2=value2", "param3=value3"))
}
