% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_paths}
\alias{count_paths}
\title{Count different paths found in URLs}
\usage{
count_paths(url, sort = FALSE, name = "n")
}
\arguments{
\item{url}{A character vector of URLs.}

\item{sort}{Logical indicating whether to sort the output by count. Defaults
to FALSE.}

\item{name}{The name of the column containing the counts. Deafults to 'n'.}
}
\value{
A tibble with each path and its count.
}
\description{
Count different paths found in URLs
}
\examples{
count_paths(c("http://example.com/index", "http://example.com/home"))
}
