% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_ports}
\alias{count_ports}
\title{Count different port numbers used in URLs}
\usage{
count_ports(url, sort = FALSE, name = "n")
}
\arguments{
\item{url}{A character vector of URLs.}

\item{sort}{Logical indicating whether to sort the output by count. Defaults
to FALSE.}

\item{name}{The name of the column containing the counts. Defaults to 'n'.}
}
\value{
A tibble with each port and how many times it occurs.
}
\description{
Count different port numbers used in URLs
}
\examples{
count_ports(c("http://example.com:8080", "http://example.com:80"))
}
