% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_queries}
\alias{count_queries}
\title{Count the occurrence of query strings in URLs}
\usage{
count_queries(url, sort = FALSE, name = "n")
}
\arguments{
\item{url}{A character vector of URLs.}

\item{sort}{Logical indicating whether to sort the output by count. Defaults
to FALSE.}

\item{name}{The name of the column containing the counts. Deafults to 'n'.}
}
\value{
A tibble with each query string and how often it occurs.
}
\description{
Count the occurrence of query strings in URLs
}
\examples{
count_queries(c("http://example.com?query1=value1", "http://example.com?query2=value2"))
}
