% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_schemes}
\alias{count_schemes}
\title{Count different schemes used in URLs}
\usage{
count_schemes(url, sort = FALSE, name = "n")
}
\arguments{
\item{url}{A character vector of URLs.}

\item{sort}{Logical indicating whether to sort the output by count. Defaults
to FALSE.}

\item{name}{The name of the column containing the counts. Defaults to 'n'.}
}
\value{
A tibble with each scheme and its count.
}
\description{
Count different schemes used in URLs
}
\examples{
count_schemes(c("http://example.com", "https://example.com"))
}
