% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_userinfos}
\alias{count_userinfos}
\title{Count occurrences of userinfo in URLs}
\usage{
count_userinfos(url, sort = FALSE, name = "n")
}
\arguments{
\item{url}{A character vector of URLs.}

\item{sort}{Logical indicating whether to sort the output by count. Defaults
to FALSE.}

\item{name}{The name of the column containing the counts. Deafults to 'n'.}
}
\value{
A tibble listing userinfos and how often each occurs.
}
\description{
Count occurrences of userinfo in URLs
}
\examples{
count_userinfos(c("user:pass@example.com", "example.com"))
}
