% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_fragment}
\alias{extract_fragment}
\title{Extract the fragment from URL}
\usage{
extract_fragment(url)
}
\arguments{
\item{url}{A character vector of URLs.}
}
\value{
A character vector containing the fragment from each URL, if present.
}
\description{
Extract the fragment from URL
}
\examples{
extract_fragment(c("http://example.com/#sec1", "http://example.com/#sec2"))
}
