% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_host}
\alias{extract_host}
\title{Extract the host from URL}
\usage{
extract_host(url)
}
\arguments{
\item{url}{A character vector of URLs.}
}
\value{
A character vector containing the host from each URL.
}
\description{
Extract the host from URL
}
\examples{
extract_host(c("https://example.com", "http://www.example.com"))
}
