% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_param_value}
\alias{extract_param_value}
\title{Extract the value of a specified parameter from the query string}
\usage{
extract_param_value(query, param_name)
}
\arguments{
\item{query}{A character vector of query strings.}

\item{param_name}{The name of the parameter to extract values for.}
}
\value{
A character vector containing the value of the specified parameter
from each query string.
}
\description{
Extract the value of a specified parameter from the query string
}
\examples{
extract_param_value(c("param1=val1&param2=val2", "param1=val3"), "param1")
}
