% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_path}
\alias{extract_path}
\title{Extract the path from URL}
\usage{
extract_path(url)
}
\arguments{
\item{url}{A character vector of URLs.}
}
\value{
A character vector containing the path from each URL.
}
\description{
Extract the path from URL
}
\examples{
extract_path(c("http://example.com/", "http://example.com/path/to/resource"))
}
