% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_path_segment}
\alias{extract_path_segment}
\title{Extract a specific segment from a path}
\usage{
extract_path_segment(path, segment_index)
}
\arguments{
\item{path}{A character vector of paths.}

\item{segment_index}{The index of the segment to extract.}
}
\value{
A character vector containing the specified segment from each path.
}
\description{
Extract a specific segment from a path
}
\examples{
extract_path_segment(c("/path/to/resource", "/another/path/"), 2)
}
