% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_query}
\alias{extract_query}
\title{Extract the query from URL}
\usage{
extract_query(url)
}
\arguments{
\item{url}{A character vector of URLs.}
}
\value{
A character vector containing the query string from each URL.
}
\description{
Extract the query from URL
}
\examples{
extract_query(c(
  "http://example.com?query1=value1&query2=value2",
  "http://example.com?query1=value3"
))
}
