% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_scheme}
\alias{extract_scheme}
\title{Extract the scheme from URL}
\usage{
extract_scheme(url)
}
\arguments{
\item{url}{A character vector of URLs.}
}
\value{
A character vector containing the scheme from each URL.
}
\description{
Extract the scheme from URL
}
\examples{
extract_scheme(c("http://example.com", "https://example.com"))
}
