% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_userinfo}
\alias{extract_userinfo}
\title{Extract userinfo from URL}
\usage{
extract_userinfo(url)
}
\arguments{
\item{url}{A character vector of URLs.}
}
\value{
A character vector containing the userinfo from each URL, if present.
}
\description{
Extract userinfo from URL
}
\examples{
extract_userinfo(c("user:pass@example.com"))
}
