% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_host}
\alias{split_host}
\title{Split host into subdomains and domain}
\usage{
split_host(host)
}
\arguments{
\item{host}{A character vector of hostnames to be split.}
}
\value{
A tibble with one row per hostname and columns for top-level domain,
domain and subdomains. Columns are created as many as the number of hosts'
components and are named as tld, domain, subdomain_1, subdomain_2, etc.
}
\description{
Split host into subdomains and domain
}
\examples{
split_host(c("subdomain.example.com"))
split_host(c("subdomain2.subdomain1.example.com", "example.com"))
}
