% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_path}
\alias{split_path}
\title{Split path into segments}
\usage{
split_path(path)
}
\arguments{
\item{path}{A character vector of paths to be split.}
}
\value{
A tibble with one row per path and columns for each segment separated
by '/'.
}
\description{
Split path into segments
}
\examples{
split_path(c("/path/to/resource"))
}
