% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_query}
\alias{split_query}
\title{Split query into parameters}
\usage{
split_query(query)
}
\arguments{
\item{query}{A character vector of query strings to be split.}
}
\value{
A tibble with one row per query string and columns for each
parameter, column names as parameter names.
}
\description{
Split query into parameters
}
\examples{
split_query(c("param1=value1&param2=value2"))
}
