% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_url}
\alias{split_url}
\title{Split URL into its constituent parts}
\usage{
split_url(url)
}
\arguments{
\item{url}{A character vector of URLs to be split.}
}
\value{
A tibble with one row per URL and columns for each component: scheme,
host, port, userinfo, path, query, and fragment.
}
\description{
Split URL into its constituent parts
}
\examples{
split_url(c("https://example.com/path?query=arg#frag"))
}
