% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlexplorer-package.R
\docType{package}
\name{urlexplorer-package}
\alias{urlexplorer}
\alias{urlexplorer-package}
\title{urlexplorer: Structural Analysis and Pattern Discovery in URL Datasets}
\description{
urlexplorer offers tools for parsing and analyzing URL datasets, extracting key components and identifying common patterns. It aids in examining website architecture and identifying SEO issues, helping users optimize web presence and content strategy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MarekProkop/urlexplorer}
  \item Report bugs at \url{https://github.com/MarekProkop/urlexplorer/issues}
}

}
\author{
\strong{Maintainer}: Marek Prokop \email{mprokop@prokopsw.cz}

}
\keyword{internal}
