% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{websitepages}
\alias{websitepages}
\title{Sample web site URLs}
\format{
\subsection{\code{websitepages}}{

A data frame with 1,000 rows and 1 column:
\describe{
\item{page}{Page URL}
...
}
}
}
\source{
Syntetic data
}
\usage{
websitepages
}
\description{
Sample web site URLs
}
\keyword{datasets}
